requirements.txt
ephem>=4.1.4
numpy>=1.20.0
matplotlib>=3.4.0
pytz>=2021.1
pillow==10.2.0
pyyaml>=6.0.1
customtkinter==5.2.2 